example/requirements.txt
-e ..
django-redis
